// prisma/generated/zod/inputTypeSchemas/OrganizationIncludeSchema.ts
import { z } from 'zod';
import type { Prisma } from '@prisma/client';
import { MembershipFindManyArgsSchema } from "../outputTypeSchemas/MembershipFindManyArgsSchema"
import { LocationFindManyArgsSchema } from "../outputTypeSchemas/LocationFindManyArgsSchema"
import { RoleFindManyArgsSchema } from "../outputTypeSchemas/RoleFindManyArgsSchema"
import { MachineFindManyArgsSchema } from "../outputTypeSchemas/MachineFindManyArgsSchema"
import { IssueFindManyArgsSchema } from "../outputTypeSchemas/IssueFindManyArgsSchema"
import { PriorityFindManyArgsSchema } from "../outputTypeSchemas/PriorityFindManyArgsSchema"
import { IssueStatusFindManyArgsSchema } from "../outputTypeSchemas/IssueStatusFindManyArgsSchema"
import { CollectionTypeFindManyArgsSchema } from "../outputTypeSchemas/CollectionTypeFindManyArgsSchema"
import { IssueHistoryFindManyArgsSchema } from "../outputTypeSchemas/IssueHistoryFindManyArgsSchema"
import { AttachmentFindManyArgsSchema } from "../outputTypeSchemas/AttachmentFindManyArgsSchema"
import { PinballMapConfigArgsSchema } from "../outputTypeSchemas/PinballMapConfigArgsSchema"
import { OrganizationCountOutputTypeArgsSchema } from "../outputTypeSchemas/OrganizationCountOutputTypeArgsSchema"

export const OrganizationIncludeSchema: z.ZodType<Prisma.OrganizationInclude> = z.object({
  memberships: z.union([z.boolean(),z.lazy(() => MembershipFindManyArgsSchema)]).optional(),
  locations: z.union([z.boolean(),z.lazy(() => LocationFindManyArgsSchema)]).optional(),
  roles: z.union([z.boolean(),z.lazy(() => RoleFindManyArgsSchema)]).optional(),
  machines: z.union([z.boolean(),z.lazy(() => MachineFindManyArgsSchema)]).optional(),
  issues: z.union([z.boolean(),z.lazy(() => IssueFindManyArgsSchema)]).optional(),
  priorities: z.union([z.boolean(),z.lazy(() => PriorityFindManyArgsSchema)]).optional(),
  issueStatuses: z.union([z.boolean(),z.lazy(() => IssueStatusFindManyArgsSchema)]).optional(),
  collectionTypes: z.union([z.boolean(),z.lazy(() => CollectionTypeFindManyArgsSchema)]).optional(),
  issueHistory: z.union([z.boolean(),z.lazy(() => IssueHistoryFindManyArgsSchema)]).optional(),
  attachments: z.union([z.boolean(),z.lazy(() => AttachmentFindManyArgsSchema)]).optional(),
  pinballMapConfig: z.union([z.boolean(),z.lazy(() => PinballMapConfigArgsSchema)]).optional(),
  _count: z.union([z.boolean(),z.lazy(() => OrganizationCountOutputTypeArgsSchema)]).optional(),
}).strict()

export default OrganizationIncludeSchema;